% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{built_in_length_hg38}
\alias{built_in_length_hg38}
\title{List of pre-computed data}
\format{
:
built_in_length_hg38: Gene length referenc for hg38.
}
\usage{
built_in_length_hg38
}
\description{
built_in_length_hg38
}
\keyword{datasets}
