% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{built_in_length_mm10}
\alias{built_in_length_mm10}
\title{List of pre-computed data}
\format{
:
built_in_length_mm10: Gene length referenc for mm10.
}
\usage{
built_in_length_mm10
}
\description{
built_in_length_mm10
}
\keyword{datasets}
