% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{built_in_lnc_ref_m39}
\alias{built_in_lnc_ref_m39}
\title{built_in_lnc_ref_m39}
\format{
:
built_in_lnc_ref_m39: reference for lncRNAs: mouse.
}
\usage{
built_in_lnc_ref_m39
}
\description{
built_in_lnc_ref_m39
}
\keyword{datasets}
