% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOnKEGG.R
\name{GOnKEGG}
\alias{GOnKEGG}
\title{Run GO and KEGG analysis}
\usage{
GOnKEGG(
  sign_de_gene,
  species = "human",
  TermsToPlot = 10,
  qcutoff = 0.01,
  pcutoff = 0.05,
  minGSSize = 10,
  maxGSSize = 500,
  Method = "BH",
  simcutoff = 0.7,
  filename = "filename"
)
}
\arguments{
\item{sign_de_gene}{A list of gene name (gene symbols).}

\item{species}{"human" or "mouse".}

\item{TermsToPlot}{Number of top enriched terms to plot. If number of enriched term < \emph{TermsToPlot}, all enriched
terms will be plotted.}

\item{qcutoff}{qvalue cutoff on enrichment tests to report as significant.}

\item{pcutoff}{Adjusted pvalue cutoff on enrichment tests to report.}

\item{minGSSize}{Minimal size of genes annotated by Ontology term for testing. Only applied to GO analysis}

\item{maxGSSize}{Minimal size of genes annotated by Ontology term for testing. Only applied to GO analysis}

\item{Method}{One of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none", method to adjust
pvalue. Only applied to GO analysis.}

\item{simcutoff}{Similarity cutoff to remove redundant GO terms. GO term with smallest padj value will be kept.}

\item{filename}{File prefix of the outputs.}
}
\description{
This function takes a lits of gene symbols (gene names) and perform GO and KEGG analysis.
}
\details{
This function create separate lists of BP, MF, CC, and KEGG enriched terms. This function also generate plots
showing the top enriched terms in each ontology (BP, MF, CC, KEGG) and a combined plot showing top enriched
GO terms (BP, MF, CC).
}
