% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergeFeatureCounts.R
\name{MergeFeatureCounts}
\alias{MergeFeatureCounts}
\title{Merge FeatureCount outputs into one table}
\usage{
MergeFeatureCounts(filepath, Assay, colDataPath, write_to_file = T)
}
\arguments{
\item{filepath}{Path to the folder storing FeatureCounts outputs. Use the "raw_counts" folder in RNAseqPE outputs.}

\item{Assay}{File prefix of the output table".}

\item{colDataPath}{File name of the Sample information text file. This file contain at least three columns:
sample_name, short_name, condition_1, condition_2, ... Sample_name should be the same as the "stat/sample,csv"
in RNAseqPE outputs.}

\item{write_to_file}{True of False, write to a file or not. If True, then use assay as file prefix.}
}
\value{
Merged FeatureCounts table.
}
\description{
This function merges FeatureCounts outputs from RNAseqPE and summarizes the results
into a text table.
}
\details{
The output table can be used for downstream DE analysis.
}
