% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseqDEGs.R
\name{RNAseqDEGs}
\alias{RNAseqDEGs}
\title{Run DEG analysis using raw counts}
\usage{
RNAseqDEGs(
  coldata,
  select1 = c(),
  counts_matrix,
  tpm_file = paste(Assay, "_tpm.csv", sep = ""),
  filename,
  LRT = FALSE,
  reduced = "",
  fitType = "parametric",
  Design = ~condition,
  shrinkcoef = 2,
  shrinktype = "apeglm",
  intgroup = "condition",
  lnc = FALSE,
  lnc_ref = "human",
  By = "padj",
  lfc = 1,
  plot_coding = FALSE,
  plimit = 1e-10,
  xlimit = 10
)
}
\arguments{
\item{select1}{A vector of short_names of samples to be included in the analysis. By defualt, all samples in the
coldata will be included.}

\item{counts_matrix}{A Merged featureCounts table generated by \code{\link[=MergeFeatureCounts]{MergeFeatureCounts()}}. \emph{colnames(counts_matrix)}
should be the same as \emph{coldata$short_name}.}

\item{tpm_file}{Path to the tpm file generated by \code{\link[=TPM_c]{TPM_c()}}.}

\item{filename}{Prefix of the outputs files.}

\item{LRT}{True of False, whether to use "LRT" test. If False, Wald test will be used.}

\item{reduced}{The reduced formula for LRT test. Only used when LRT = T.}

\item{fitType}{Either "parametric", "local", "mean", or "glmGamPoi" for the type of fitting of dispersions
to the mean intensity. see \code{\link[=DESeq]{DESeq()}} for details.}

\item{Design}{The formula or matrix passed to DESeq2. The formula expresses how the counts for each gene depend
on the variables in colData File prefix of the outputs".}

\item{shrinkcoef}{The name or number of the coefficient (LFC) to shrink. See \code{\link[=lfcShrink]{lfcShrink()}}}

\item{shrinktype}{Type of Shrink to perform, See \code{\link[=lfcShrink]{lfcShrink()}}}

\item{intgroup}{The column in coldata for coloring samples in the PCA plots}

\item{lnc}{True of False, whether to separate non coding RNAs versus coding RNAs.}

\item{lnc_ref}{Refernce for non coding RNAs. It can be either "human", "mouse", or a user-provided
vector containing gene names of ncRNAs.}

\item{By}{"padj" or "pvalue" to use.}

\item{lfc}{Threshold for |log2FC| when ploting.}

\item{plot_coding}{True of False, whether to only include coding RNAs in the volcano plot.}

\item{plimit}{Limitation of Pvalue/Padj to show in the plots. Pvalue < plimit will be convert to plimit.}

\item{xlimit}{Limitation of |log2FC| to show in the plots. |log2FC| > xlimit will be convert to xlimit.}

\item{colData}{Sample information text file. This file contain at least three columns:
sample_name, short_name, condition_1, condition_2, ... sample_name should be the same as the "stat/sample,csv"
in RNAseqPE outputs. Only used if Plots = T}
}
\value{
A list of three objects: a sample information matrix, dds, and \emph{result(dds)}.
}
\description{
This function takes raw count table and run the full DESeq2 analysis to identify DEGs.
}
\details{
This function create a folder containing all the output files.
}
