% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPM_c.R
\name{TPM_c}
\alias{TPM_c}
\title{Convert FeatureCounts to TPM (transcript per million nucleotide)}
\usage{
TPM_c(
  Assay = Assay,
  counts_matrix = counts_matrix,
  lengthdata = "hg38",
  Plots = T,
  coldata = coldata,
  pca_colors = paste(coldata$condition)
)
}
\arguments{
\item{Assay}{File prefix of the outputs".}

\item{counts_matrix}{A Merged featureCounts table generated by \code{\link[=MergeFeatureCounts]{MergeFeatureCounts()}}.}

\item{lengthdata}{Data.frame containing the gene length. It could be either "hg38", "mm10", or any customized
data.frame. Customized data.frame should have gene names as rownames and only one column containg gene length.}

\item{Plots}{True of False, whether to plot sample correlations. If Ture, heatmap and PCA plots will be generated
with the same file prefix.}

\item{coldata}{Sample information text file. This file contain at least three columns:
sample_name, short_name, condition_1, condition_2, ... sample_name should be the same as the "stat/sample,csv"
in RNAseqPE outputs. Only used if Plots = T}

\item{pca_colors}{Sample groups used to color the PCA points. This should be a vector of factors same length as
\link{ncol(counts_matrix)}. Default is the \link{ColData$conditions}. Only used if Plots = T.}
}
\description{
This function converts a Merged featureCounts table into TPM table. Merged featureCounts table is generated by
\code{\link[=MergeFeatureCounts]{MergeFeatureCounts()}}.
}
\details{
This function create two files: a table containing CPM in each sample
and a table containing average CPM among all samples. If Plots == T, it also plot a correlation heatmap and a
PCA plot based on the TPM
}
